Restructured the database directory for MySQL specific files. Added a clean slate database with table creation already pre-defined. cardiathena.sql pre-populates the agents table with The Random Agent and The Low Layer Agent, and will need to get updated with additional agents.
