+++ database/MySql/MySQLVariables.py
